remove in-process grappler in favor of commandGrappler
